+++ e2e-tests/standalone-parser/src/parser.ts
@@ -1,6 +1,8 @@
 import type { DeploymentState } from "./decl";
 import parse from "./generated/parser";
 import { Arr2, OtherEnum, ValidCurrency, OtherEnum2, Arr3 } from "./types";
+import { TsxTask } from "./types-tsx";
+import type { DtsConfig } from "./types-dts";
 import {
   CurrencyPrices,
   NonInfiniteNumber,
@@ -262,6 +264,8 @@ export const {
   Int32ArrayCodec: Int32ArrayCodec,
   Float64ArrayCodec: Float64ArrayCodec,
   BigInt64ArrayCodec: BigInt64ArrayCodec,
+  TsxTask: TsxTaskCodec,
+  DtsConfig: DtsConfigCodec,
 } = parse.buildParsers<{
   PartialRepro: PartialRepro;
   TransportedValue: TransportedValue;
@@ -322,6 +326,8 @@ export const {
   Int32ArrayCodec: Int32ArrayType;
   Float64ArrayCodec: Float64ArrayType;
   BigInt64ArrayCodec: BigInt64ArrayType;
+  TsxTask: TsxTask;
+  DtsConfig: DtsConfig;
   DeploymentState: DeploymentState;
 }>({
   stringFormats: {