+++ packages/beff-client/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @beff/client
 
+## 0.0.124
+
+### Patch Changes
+
+- parse tsx files
+- Updated dependencies
+  - @beff/cli@0.0.124
+
 ## 0.0.123
 
 ### Patch Changes